Remove SearchPhase interface
